pull changes

+++ ZarvanOrder/Services/Data/CustomerService.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using ZarvanOrder.CustomException;
 using ZarvanOrder.Interfaces.DataServices;
 using ZarvanOrder.Interfaces.Repositores;
 using ZarvanOrder.Model.Dtos.Requests.Customers;
@@ -20,10 +21,10 @@ namespace ZarvanOrder.Services.Data
     public class CustomerService : ICustomerService
     {
         private readonly ICustomerRepository _repository;
-        private readonly Mapper _mapper;
+        private readonly IMapper _mapper;
 
         public CustomerService(ICustomerRepository repository,
-                               Mapper mapper)
+                               IMapper mapper)
         {
             this._repository = repository;
             this._mapper = mapper;
@@ -61,7 +62,9 @@ namespace ZarvanOrder.Services.Data
 
         public async Task Delete(DeleteCustomerRequest request)
         {
-            Customer entity = new();
+            Customer entity = await _repository.GetById(_mapper.Map<GetCustomerRequest>(request));
+            if (entity == null)
+                throw new MyException(System.Net.HttpStatusCode.NotFound, Model.Messages.General.CustomerNotFound);
             _mapper.Map<DeleteCustomerRequest, Customer>(request, entity);
             var result = _mapper.Map<CustomerResponse>(_repository.Delete(entity));
             await Task.Run(() => result);
@@ -77,15 +80,11 @@ namespace ZarvanOrder.Services.Data
             var items = _mapper.Map<List<CustomerResponse>>(await _repository.Get(request).ToListAsync());
             return new ListResponse<CustomerResponse>() { Items = items, Total = items.Count() };
         }
-
-        public Task<bool> IsUniqueCustomerCode(string Code)
-        {
-            _repository.Get(new GetCustomersRequest() { Code = Code })
-        }
-
         public async Task<CustomerResponse> Update(EditCustomerRequest request)
         {
-            Customer entity = new();
+            Customer entity = await _repository.GetById(_mapper.Map<GetCustomerRequest>(request));
+            if (entity == null)
+                throw new MyException(System.Net.HttpStatusCode.NotFound, Model.Messages.General.CustomerNotFound);
             _mapper.Map<EditCustomerRequest, Customer>(request, entity);
             CustomerValidation validator = new(this);
             await validator.ValidateAndThrowAsync(entity);
